+++ core-processor/src/main/java/jeff/core/config/DevEnvConfig.java
@@ -11,8 +11,11 @@ import org.springframework.context.annotation.PropertySources;
 @Configuration
 @Profile("dev")
 @PropertySources({
-        @PropertySource({"classpath:myDataSource-dev.properties",
-                "classpath:myRedis-dev.properties"})})
+        @PropertySource({
+                "classpath:myDataSource-dev.properties",
+                "classpath:myRedis-dev.properties",
+                "classpath:myMQ-dev.properties"
+        })})
 public class DevEnvConfig {
 
 }