+++ core-processor/src/main/java/jeff/core/filter/UUIDFilter.java
@@ -21,7 +21,7 @@ import java.io.IOException;
  */
 @Component
 @WebFilter
-@Order(5)
+@Order(FilterOrderNumber.UUID_FILTER)
 public class UUIDFilter extends OncePerRequestFilter {
 
     @Autowired